+++ src/main/java/com/example/Proveedores_Empresariales/BranchOfficeCompan/BranchOfficeCompanService.java
@@ -3,6 +3,7 @@ package com.example.Proveedores_Empresariales.BranchOfficeCompan;
 import com.example.Proveedores_Empresariales.City.City;
 import org.springframework.stereotype.Service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Service
@@ -12,9 +13,9 @@ public interface BranchOfficeCompanService {
 
     public List<BranchOfficeCompan> getAll();
 
-    public BranchOfficeCompan getById(int id);
+    public BranchOfficeCompan getById(BigInteger id);
 
-    public BranchOfficeCompan update(int nic,BranchOfficeCompan branchOfficeCompan);
+    public BranchOfficeCompan update(BigInteger nic,BranchOfficeCompan branchOfficeCompan);
 
     public void delete(BranchOfficeCompan branchOfficeCompan);
 }
