+++ src/main/java/com/example/Proveedores_Empresariales/BranchOfficeCompan/RepositoryBranchOfficeCompan.java
@@ -2,8 +2,14 @@ package com.example.Proveedores_Empresariales.BranchOfficeCompan;
 
 import com.example.Proveedores_Empresariales.City.City;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface RepositoryBranchOfficeCompan extends JpaRepository<BranchOfficeCompan,Integer> {
+public interface RepositoryBranchOfficeCompan extends JpaRepository<BranchOfficeCompan, BigInteger> {
+
+
 }
